Move layout context to SidabarPage
